ejemplo/pyproject.toml
[tool.poetry]
name = "django-payments-chile-ejemplo"
version = "0.1.0"
description = ""
authors = ["Mario Hernandez <mariofix@proton.me>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django-payments-chile = "^2024.9.29"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
